完善用户添加和修改接口

+++ bmss/src/main/java/com/qsz/bmss/dao/SystemRoleDao.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 @Repository
 public interface SystemRoleDao extends BaseMapper<SystemRole> {
-    List<SystemRole> selectRolesByMenuId(Long menuId);
+    List<SystemRole> selectRolesByMenuId(Integer menuId);
 }
