+++ bmss/src/main/java/com/qsz/bmss/domain/SystemUser.java
@@ -4,9 +4,12 @@ import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import java.beans.Transient;
+import java.util.Date;
 import java.util.List;
 
 
@@ -15,7 +18,7 @@ import java.util.List;
 public class SystemUser {
 
     @TableId
-    private Long userId;
+    private Integer userId;
 
     @TableField("user_name")
     private String username;
