+++ bmss/src/main/java/com/qsz/bmss/security/utils/SecurityUtil.java
@@ -22,7 +22,7 @@ public class SecurityUtil {
     /**
      * 获取当前用户ID
      */
-    public static Long getUserId(){
+    public static Integer getUserId(){
         return getUserInfo().getUserId();
     }
     /**
