+++ bmss/src/main/java/com/qsz/bmss/service/ISystemRoleService.java
@@ -5,7 +5,7 @@ import com.qsz.bmss.domain.SystemRole;
 import java.util.List;
 
 public interface ISystemRoleService {
-    List<SystemRole> selectRolesByMenuId(Long menuId);
+    List<SystemRole> selectRolesByMenuId(Integer menuId);
 
     List<SystemRole> getAllRoles();
 }
